+++ scripts/Tool/jass_tool.lua
@@ -5,6 +5,8 @@ return {
     H2S = function(h) return cj.I2S(H2I(h)) end,
     U2PlayerId = function(u) return cj.GetPlayerId(cj.GetOwningPlayer(u)) end,
     Debug = function(s) cj.DisplayTimedTextToPlayer(cj.Player(0), 0, 0, 5, s) end,
-    U2Id = function(u) return cj.GetUnitTypeId(u) end
+    U2Id = function(u) return cj.GetUnitTypeId(u) end,
+    RemoveUnit = function(unit) cj.UnitApplyTimedLife(unit, Base.String2Id('BHwe'), 0.1) end
+    -- 設定生命週期利用war3機制自動刪除，會比用RemoveUnit乾淨，內存絕不會漏掉
 }